validation: refactor uniqueness check
